requirements.txt
fastapi>=0.95.0
uvicorn>=0.20.0
web3>=6.0.0
pydantic>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
pandas>=2.0.0
torch>=2.0.0
transformers>=4.30.0
pynacl>=1.5.0
ipfshttpclient>=0.8.0
eth-account>=0.10.0